# speed up marshaling of transport parameters