This is not in the critical path, so technically this is not a part of #3526. It's fun to see how much we can gain anyway, and such an easy and self-contained change that it's definitely worth doing.

The speedup comes from multiple sources:
1. We now preallocate a byte slice, instead of appending multiple times.
2. Marshaling into a byte slice is faster than using a bytes.Buffer.
3. quicvarint.Write allocates, while quicvarint.Append doesn't.

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkTransportParametersMarshal-10     368           196           -46.80%

benchmark                                  old allocs     new allocs     delta
BenchmarkTransportParametersMarshal-10     14             1              -92.86%

benchmark                                  old bytes     new bytes     delta
BenchmarkTransportParametersMarshal-10     273           256           -6.23%
```